Make ping tests work again
