+++ lib/src/core/network/wiredash_api.dart
@@ -5,12 +5,14 @@ import 'dart:typed_data';
 
 import 'package:http/http.dart';
 import 'package:http_parser/http_parser.dart';
+import 'package:wiredash/src/core/network/api_exceptions.dart';
 import 'package:wiredash/src/core/network/ping_request.dart';
 import 'package:wiredash/src/core/network/send_feedback_request.dart';
 import 'package:wiredash/src/core/network/send_promoter_score_request.dart';
 import 'package:wiredash/src/core/network/upload_attachment_request.dart';
 import 'package:wiredash/src/feedback/data/feedback_item.dart';
 
+export 'package:wiredash/src/core/network/api_exceptions.dart';
 export 'package:wiredash/src/core/network/ping_request.dart';
 export 'package:wiredash/src/core/network/send_feedback_request.dart';
 export 'package:wiredash/src/core/network/send_promoter_score_request.dart';
@@ -97,85 +99,5 @@ class ApiClientContext {
       throw KillSwitchException(response: response);
     }
     throw WiredashApiException(response: response);
-  }
-}
-
-/// Generic error from the Wiredash API
-class WiredashApiException implements Exception {
-  const WiredashApiException({this.message, this.response});
-
-  String? get messageFromServer {
-    try {
-      // Official error format for wiredash backend
-      final json = jsonDecode(response?.body ?? '') as Map?;
-      final message = json?['errorMessage'] as String?;
-      final code = json?['errorCode'] as int?;
-      if (code != null) {
-        return '[$code] ${message ?? '<no message>'}';
-      }
-      return message!;
-    } catch (_) {
-      // ignore
-    }
-    try {
-      // Parsing errors often have this format
-      final json = jsonDecode(response?.body ?? '') as Map?;
-      final message = json?['message'] as String?;
-      return message!;
-    } catch (_) {
-      // ignore
-    }
-    return response?.body;
-  }
-
-  final String? message;
-  final Response? response;
-
-  @override
-  String toString() {
-    return 'WiredashApiException{'
-        '"$message", '
-        'endpoint: ${response?.request?.url.path}, '
-        'code: ${response?.statusCode}, '
-        'resp: $messageFromServer'
-        '}';
-  }
-}
-
-/// Thrown when the server couldn't match the project + secret to a existing
-/// project
-class UnauthenticatedWiredashApiException extends WiredashApiException {
-  UnauthenticatedWiredashApiException(
-    Response response,
-    this.projectId,
-    this.secret,
-  ) : super(
-          message: "Unknown projectId:'$projectId' or invalid secret:'$secret'",
-          response: response,
-        );
-
-  final String projectId;
-  final String secret;
-
-  @override
-  String toString() {
-    return 'UnauthenticatedWiredashApiException{'
-        '$message, '
-        'status code: ${response?.statusCode}, '
-        '${response?.body}'
-        '}';
-  }
-}
-
-enum AttachmentType {
-  screenshot,
-}
-
-/// Backend returns an error which silences the SDK for one week
-class KillSwitchException extends WiredashApiException {
-  const KillSwitchException({super.response});
-  @override
-  String toString() {
-    return 'KillSwitchException{${response?.statusCode}, body: ${response?.body}}';
   }
 }
